[fix] Branding API findAllBrandingListByDesignerId error Handling 구문 추가
